+++ CHANGELOG.md
@@ -98,6 +98,10 @@ edits the change twice in some cases.
 * Git-based repositories are now colocated by default. Configure `git.colocate =
   false` to keep the previous behavior.
 
+* The new `jj git colocation` command provides sub-commands to show the
+  colocation state (`status`) and to convert a non-colocated git repo into
+  a colocated repo (`enable`) and vice-versa `disable`.
+
 * Conflicts written by jj < 0.11 are no longer supported. They will now appear
   as regular files with a `.jjconflict` suffix and JSON contents.